// src/types/financial.ts
export interface DuPontAnalysis {
  roe: number;
  netProfitMargin: number;
  assetTurnover: number;
  financialLeverage: number;
  npmContribution: number;
  atoContribution: number;
  flContribution: number;
}